requirements.txt
albumentations==1.1.0
imutils==0.5.4
matplotlib==3.5.1
numpy==1.22.3
opencv_python_headless==4.5.5.64
Pillow==9.1.0
scikit_learn==1.0.2
tensorflow==2.8.0
tensorflow_gpu==2.8.0
tqdm==4.64.0
